+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.ClientModel/src/Providers/MultiPartFormDataBinaryContentDefinition.cs
@@ -80,7 +80,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
 
         protected override TypeSignatureModifiers BuildDeclarationModifiers() => TypeSignatureModifiers.Class | TypeSignatureModifiers.Internal;
 
-        protected override CSharpType? GetBaseType() => typeof(BinaryContent);
+        protected override CSharpType? BuildBaseType() => typeof(BinaryContent);
 
         protected override string BuildRelativeFilePath() => Path.Combine("src", "Generated", "Internal", $"{Name}.cs");
 
@@ -97,10 +97,10 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
         protected override ConstructorProvider[] BuildConstructors()
         {
             var signature = new ConstructorSignature(
-                Type: Type,
-                Modifiers: MethodSignatureModifiers.Public,
-                Parameters: [],
-                Description: null);
+                type: Type,
+                modifiers: MethodSignatureModifiers.Public,
+                parameters: [],
+                description: null);
 
             return
             [