+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.ClientModel/test/Providers/ClientProviders/TestData/ClientProviderTests/TestBuildOAuth2FlowsField(NoFlows).cs
@@ -1,0 +1,1 @@
+Array.Empty<global::System.Collections.Generic.Dictionary<string, object>>()