+++ packages/http-client-java/generator/http-client-generator-test/src/main/java/tsptest/armresourceprovider/implementation/ManagedMaintenanceWindowStatusOperationsClientImpl.java
@@ -87,7 +87,7 @@ public final class ManagedMaintenanceWindowStatusOperationsClientImpl
             @PathParam("managedMaintenanceWindowStatusContentName") String managedMaintenanceWindowStatusContentName,
             @HeaderParam("Accept") String accept, Context context);
 
-        @Headers({ "Content-Type: application/json" })
+        @Headers({ "Accept: application/json;q=0.9", "Content-Type: application/json" })
         @Delete("/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/TspTest.ArmResourceProvider/managedMaintenanceWindowStatusContents/{managedMaintenanceWindowStatusContentName}")
         @ExpectedResponses({ 200, 202, 204 })
         @UnexpectedResponseExceptionType(ManagementException.class)
@@ -95,10 +95,9 @@ public final class ManagedMaintenanceWindowStatusOperationsClientImpl
             @QueryParam("api-version") String apiVersion, @PathParam("subscriptionId") String subscriptionId,
             @PathParam("resourceGroupName") String resourceGroupName,
             @PathParam("managedMaintenanceWindowStatusContentName") String managedMaintenanceWindowStatusContentName,
-            @HeaderParam("If-Match") String ifMatch, @HeaderParam("If-None-Match") String ifNoneMatch,
-            @HeaderParam("Accept") String accept, Context context);
+            @HeaderParam("If-Match") String ifMatch, @HeaderParam("If-None-Match") String ifNoneMatch, Context context);
 
-        @Headers({ "Content-Type: application/json" })
+        @Headers({ "Accept: application/json;q=0.9", "Content-Type: application/json" })
         @Delete("/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/TspTest.ArmResourceProvider/managedMaintenanceWindowStatusContents/{managedMaintenanceWindowStatusContentName}")
         @ExpectedResponses({ 200, 202, 204 })
         @UnexpectedResponseExceptionType(ManagementException.class)
@@ -106,8 +105,7 @@ public final class ManagedMaintenanceWindowStatusOperationsClientImpl
             @QueryParam("api-version") String apiVersion, @PathParam("subscriptionId") String subscriptionId,
             @PathParam("resourceGroupName") String resourceGroupName,
             @PathParam("managedMaintenanceWindowStatusContentName") String managedMaintenanceWindowStatusContentName,
-            @HeaderParam("If-Match") String ifMatch, @HeaderParam("If-None-Match") String ifNoneMatch,
-            @HeaderParam("Accept") String accept, Context context);
+            @HeaderParam("If-Match") String ifMatch, @HeaderParam("If-None-Match") String ifNoneMatch, Context context);
     }
 
     /**
@@ -252,11 +250,10 @@ public final class ManagedMaintenanceWindowStatusOperationsClientImpl
             return Mono.error(new IllegalArgumentException(
                 "Parameter managedMaintenanceWindowStatusContentName is required and cannot be null."));
         }
-        final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.delete(this.client.getEndpoint(), this.client.getApiVersion(),
                 this.client.getSubscriptionId(), resourceGroupName, managedMaintenanceWindowStatusContentName, ifMatch,
-                ifNoneMatch, accept, context))
+                ifNoneMatch, context))
             .contextWrite(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext()).readOnly()));
     }
 
@@ -294,10 +291,9 @@ public final class ManagedMaintenanceWindowStatusOperationsClientImpl
                 .log(new IllegalArgumentException(
                     "Parameter managedMaintenanceWindowStatusContentName is required and cannot be null."));
         }
-        final String accept = "application/json";
         return service.deleteSync(this.client.getEndpoint(), this.client.getApiVersion(),
             this.client.getSubscriptionId(), resourceGroupName, managedMaintenanceWindowStatusContentName, ifMatch,
-            ifNoneMatch, accept, Context.NONE);
+            ifNoneMatch, Context.NONE);
     }
 
     /**
@@ -335,10 +331,9 @@ public final class ManagedMaintenanceWindowStatusOperationsClientImpl
                 .log(new IllegalArgumentException(
                     "Parameter managedMaintenanceWindowStatusContentName is required and cannot be null."));
         }
-        final String accept = "application/json";
         return service.deleteSync(this.client.getEndpoint(), this.client.getApiVersion(),
             this.client.getSubscriptionId(), resourceGroupName, managedMaintenanceWindowStatusContentName, ifMatch,
-            ifNoneMatch, accept, context);
+            ifNoneMatch, context);
     }
 
     /**