+++ packages/http-client-python/eng/scripts/ci/regenerate.ts
@@ -58,6 +58,9 @@ const AZURE_EMITTER_OPTIONS: Record<string, Record<string, string> | Record<stri
   "azure/client-generator-core/usage": {
     namespace: "specs.azure.clientgenerator.core.usage",
   },
+  "azure/client-generator-core/override": {
+    namespace: "specs.azure.clientgenerator.core.override",
+  },
   "azure/core/basic": {
     namespace: "specs.azure.core.basic",
   },
@@ -150,10 +153,17 @@ const EMITTER_OPTIONS: Record<string, Record<string, string> | Record<string, st
     namespace: "authentication.http.custom",
     "package-pprint-name": "Authentication Http Custom",
   },
-  "authentication/union": {
-    "package-name": "authentication-union",
-    namespace: "authentication.union",
-  },
+  "authentication/union": [
+    {
+      "package-name": "authentication-union",
+      namespace: "authentication.union",
+    },
+    {
+      "package-name": "setuppy-authentication-union",
+      namespace: "setuppy.authentication.union",
+      "keep-setup-py": "true",
+    },
+  ],
   "type/array": {
     "package-name": "typetest-array",
     namespace: "typetest.array",
@@ -305,6 +315,9 @@ async function getSubdirectories(baseDir: string, flags: RegenerateFlags): Promi
         const clientTspPath = join(subDirPath, "client.tsp");
 
         const mainTspRelativePath = toPosix(relative(baseDir, mainTspPath));
+
+        // after support discriminated union, remove this check
+        if (mainTspRelativePath.includes("type/union/discriminated")) return;
 
         // after fix test generation for nested operation group, remove this check
         if (mainTspRelativePath.includes("client-operation-group")) return;