+++ packages/http-client-python/generator/pygen/codegen/serializers/model_serializer.py
@@ -264,7 +264,7 @@ class DpgModelSerializer(_ModelSerializer):
                         parent.name,
                         ImportType.LOCAL,
                     )
-            if not model.internal and self.init_line(model):
+            if self.need_init(model):
                 file_import.add_submodule_import("typing", "overload", ImportType.STDLIB)
                 file_import.add_submodule_import("typing", "Mapping", ImportType.STDLIB)
                 file_import.add_submodule_import("typing", "Any", ImportType.STDLIB)