+++ packages/json-schema/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log - @typespec/json-schema
 
+## 1.2.0
+
+### Bump dependencies
+
+- [#7674](https://github.com/microsoft/typespec/pull/7674) Upgrade dependencies
+
+
 ## 1.1.0
 
 ### Bug Fixes